+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.0]
+
+* Simplify obtaining all locales and native locale names by encoding them in Dart instead of JSON
+  * No more need for a class instance or to wait for promises
+
 ## [1.0.4]
 
 * Fix bug where native locale names could not be obtained on web if using [peanut](https://pub.dev/packages/peanut)